+++ xs/sources/xlil/parser.c
@@ -322,6 +322,9 @@ static XsLilStatus parse_binary_integer(Parser *parser, XsLilBlock *block, const
   case XS_LIL_INSTRUCTION_DIV_I32:
     status = xs_lil_block_div_i32(block, left, right, &result, error);
     break;
+  case XS_LIL_INSTRUCTION_REM_I32:
+    status = xs_lil_block_rem_i32(block, left, right, &result, error);
+    break;
   case XS_LIL_INSTRUCTION_EQ_I32:
     status = xs_lil_block_eq_i32(block, left, right, &result, error);
     break;
@@ -380,6 +383,7 @@ static XsLilStatus parse_instruction(Parser *parser, XsLilBlock *block, const ch
         {"sub.i32 ", XS_LIL_INSTRUCTION_SUB_I32, XS_LIL_TYPE_I32},
         {"mul.i32 ", XS_LIL_INSTRUCTION_MUL_I32, XS_LIL_TYPE_I32},
         {"div.i32 ", XS_LIL_INSTRUCTION_DIV_I32, XS_LIL_TYPE_I32},
+        {"rem.i32 ", XS_LIL_INSTRUCTION_REM_I32, XS_LIL_TYPE_I32},
         {"eq.i32 ", XS_LIL_INSTRUCTION_EQ_I32, XS_LIL_TYPE_BOOL},
         {"lt.i32 ", XS_LIL_INSTRUCTION_LT_I32, XS_LIL_TYPE_BOOL},
         {"le.i32 ", XS_LIL_INSTRUCTION_LE_I32, XS_LIL_TYPE_BOOL},